+++ rush.json
@@ -75,8 +75,8 @@
       "versionPolicyName": "tools"
     },
     {
-      "packageName": "@lukeshay/rollup-config-typescript",
-      "projectFolder": "tools/rollup-config-typescript",
+      "packageName": "@lukeshay/rollup-config",
+      "projectFolder": "tools/rollup-config",
       "reviewCategory": "tools",
       "shouldPublish": true,
       "versionPolicyName": "tools"
