+++ tools/eslint-config-typescript/package.json
@@ -36,5 +36,13 @@
   },
   "peerDependencies": {
     "typescript": "*"
-  }
+  },
+  "files": [
+    "index.js",
+    "CHANGELOG.md",
+    "LICENSE",
+    "README.md",
+    "package.json",
+    "patch.js"
+  ]
 }
