+++ tools/eslint-config/package.json
@@ -27,5 +27,13 @@
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.2"
   },
-  "peerDependencies": {}
+  "peerDependencies": {},
+  "files": [
+    "index.js",
+    "CHANGELOG.md",
+    "LICENSE",
+    "README.md",
+    "package.json",
+    "patch.js"
+  ]
 }
