+++ tools/rollup-config/.eslintrc.js
