+++ tools/rollup-config/README.md
@@ -1,0 +1,1 @@
+# @lukeshay/rollup-config-typescript
