+++ tools/rollup-config/rollup.config.js
@@ -1,0 +1,4 @@
+import { javascriptConfig } from './index';
+import pkg from './package.json';
+
+export default javascriptConfig(pkg);
