+++ tools/typescript-config/package.json
@@ -26,6 +26,9 @@
   "homepage": "https://github.com/LukeShay/lukeshay-web#readme",
   "files": [
     "tsconfig.json",
-    "tsconfig-web.json"
+    "tsconfig-web.json",
+    "LICENSE",
+    "package.json",
+    "CHANGELOG.md"
   ]
 }
